update client ttl type

+++ src/client/tablet_client.h
@@ -26,10 +26,10 @@ public:
     bool CreateTable(const std::string& name, 
                      uint32_t id, 
                      uint32_t pid,
-                     uint32_t ttl);
+                     uint64_t ttl);
 
     bool CreateTable(const std::string& name,
-                     uint32_t tid, uint32_t pid, uint32_t ttl,
+                     uint32_t tid, uint32_t pid, uint64_t ttl,
                      bool leader, const std::vector<std::string>& endpoints);
 
     bool Put(uint32_t tid,
@@ -73,9 +73,9 @@ public:
 
     bool LoadSnapshot(uint32_t tid, uint32_t pid);
 
-    bool LoadTable(const std::string& name, uint32_t id, uint32_t pid, uint32_t ttl);
+    bool LoadTable(const std::string& name, uint32_t id, uint32_t pid, uint64_t ttl);
 
-    bool LoadTable(const std::string& name, uint32_t id, uint32_t pid, uint32_t ttl,
+    bool LoadTable(const std::string& name, uint32_t id, uint32_t pid, uint64_t ttl,
                bool leader, const std::vector<std::string>& endpoints);
 
     bool ChangeRole(uint32_t tid, uint32_t pid, bool leader);
